Uses decoder instead of JSON gem
Okay, so apparently aws-sdk has a dependancy of the `json` gem, so
technically we could use it here.  But since the docs are expecting the
use of the MultiJson gem when encoding and decoding, it makes sense that
we use it in those places.

We are rescuing ::JSON::ParserError in the #raw_body_with_sns_check
method, but this is because the AWS object can possibly return that when
using the json gem that it requires.  For that reason, it is going to
remain in the code.